Removed redundant active player index
i and activePlayerIndex used to need to be separate variables, but due to a control flow change I made earlier, they ended up being the same variable. activePlayerIndex is more descriptive, so that is the one I kept. I removed the use of "i" in the sense of tracking the active player.