Add comments to pokerPlayer.py
Commented pokerPlayer.py
NOTE: renamed "temp" to "vals" in the method handScore(), as that is more descriptive of its function.
Moved the position of some of the scoring checks to be more intuitive (put the straight checks together).
NOTE: THE STRAIGHT LOGIC CURRENTLY DOES NOT WORK PROPERLY. Aces should NOT act as a bridge between King and Two, Aces can act as a 1 OR a 14, not both. This is not implemented properly, and should be changed.