Update poker.py with correct Check and Call rules
Calling now counts as a passive action.
Checking can now only be performed when: The incoming bet is 0 AND the checkFlag is true.
Performing any non-passive action while the checkFlag is true (anything except call, check, and fold), results in the checkFlag being set to false, and all playersPassing elements being set to false (giving checking players an opportunity to play).
The checkFlag is set to true at the beginning of each phase.