Scoring Fix and Repair
Now finds full houses when two instances of 3 of a kind are detected.
Now scores full houses, two pairs, and flushes with decreasing powers of 100 rather than 10, to eliminate scoring inaccuracies, as well as improving score visibility (the value of the subsequent cards can be seen clearly in 2 character chunks after the decimal point).
Fixes #49