testPoker creation and surface Poker changes
Poker now takes an array of players instead of a number of players, poker also now takes an array of decks as deckSequences, and handles it accordingly. WARNING: if you try to run testPoker.py when there is an uncommented running of testGame in the poker.py file, it will run that when it imports, so comment it out before you run tests.