EXPAND implementation plus button tracking in board
Added buttonPlayerIndex to board. This was done with the sole purpose of letting the agent simulations & tree know who gets to go first each betting round/phase.
selectRecursion can now return a chosen with no available actions if it has no children. This should ONLY occur at end states, where the phase incremented to 5, as the getValidPlayerActions in the MonteCarloAgent.MCTree now returns an empty array if phase is 5. EXPAND was modified to interpret this and handle it accordingly.
Copied and converted the "handle<action>" methods from round into something usable by both the tree-building component and simulation component of the monte carlo agent.
EXPAND now handles the randomly chosen player action accordingly, and does a phase update check at the end to help maintain state. Additionally, it detects if the active player is all in or folding, and essentially skips their turn as it should.
SIMULATE has some added notes about expected implementation. The highlight is that community cards will not be revealed as the tree is built. This is because doing so would limit simulations to treat these random reveals as "true game state" when they were just generated randomly in our tree. It is the responsibility of the simulator to populate the simulation board community cards with the correct amount of cards at the start of the simulation by looking at the phase of the board. Details on how to do this are included in the comment.
BACK_PROPAGATE now correctly negates the result as it alternates between Agent nodes and Opponent nodes.
I wrote the current implementation of MCNode... and I'm not sure what all the isAgent field is doing. I'm not removing it yet, since it could be useful for debugging to identify which nodes are results of decisions by which players.
I'm on the fence about just removing fold from the tree... If we do this, we'd have to implement some kind of minimum utility threshold for folding, where if, after all simulations have been run, all available options have utility scores below the threshold, we just fold instead. The immediate problem with that is apparently MCTS chooses the action with the most simulations to choose, NOT the highest utility??? we could try ours out by choosing utility and going about it that way... but I have a feeling that just leaving fold in, and using a net chip gain approach to utility would be better in the long run, and more intuitive to code. This way folding will consistently generate a small net loss, while other options that could be bad for us in a given spot would generate larger net losses. Ideally, if we have a bad hand, all options would generate net chip losses, and fold would generate the smallest chip loss.