Added Simple Playout Policies. Added U(n) modes. Added test sequences files. Allowed for variable sized hands in handScore() methods.
Added `getPlayoutPolicyActionSimple` and `getPlayoutPolicyActionSimpleProbability`. Both of which take the active player's hand and community cards into account to assign a handRank, and the incoming bet to assign a modifier. These are combined to result in a policyRank. In the former, the policyRank corresponds to an action to be performed. In the latter, the rank responds to a probability distribution of actions, where one is chosen at random from that distribution.

Utility (U(n)) now has some modes (the active one is the only one that should be uncommented). I tried a few, but the one that worked the best was the sum of the square root of the chips won(+)/lost(-). This one also performed better when the value of c in UCB1 is multiplied by 2 to give you 2sqrt(2) or sqrt(8).
Test sequences 2, 3, and 4 are 10 games long with 100-200 rounds worth of test material each. Test  sequence 5 is 25 games long with I believe 300 rounds of material.